+++ Assets/03Data/PlantData.cs
@@ -7,15 +7,15 @@ public class PlantInfo
 {
     public int id;
     public string name;
-    public int health;         // 체력
-    public int damage;         // 공격력
+    public float health;         // 체력
+    public float damage;         // 공격력
     public float attackSpeed;  // 공격 주기(초)
-    public int sunCost;        // 심는 데 필요한 태양 개수
+    public float sunCost;        // 심는 데 필요한 태양 개수
     public float rechargeTime; // 재사용 대기시간(초)
     public bool canAttack;     // 호두,체리폭탄용
 
 
-    public PlantInfo(int id, string name, int health, int damage, float attackSpeed, int sunCost, float rechargeTime, bool canAttack)
+    public PlantInfo(int id, string name, float health, float damage, float attackSpeed, float sunCost, float rechargeTime, bool canAttack)
     {
         this.id = id;
         this.name = name;
